+++ adapters/risc0/groth16-verifier/Cargo.toml
@@ -3,25 +3,22 @@ edition = "2021"
 name = "zkaleido-risc0-groth16-verifier"
 version = "0.1.0"
 
-# [lints]
-# rust.missing_debug_implementations = "warn"
-# rust.missing_docs = "warn"
-# rust.rust_2018_idioms = { level = "deny", priority = -1 }
-# rust.unreachable_pub = "warn"
-# rust.unused_crate_dependencies = "deny"
-# rust.unused_must_use = "deny"
-# rustdoc.all = "warn"
+[lints]
+rust.missing_debug_implementations = "warn"
+rust.missing_docs = "warn"
+rust.rust_2018_idioms = { level = "deny", priority = -1 }
+rust.unreachable_pub = "warn"
+rust.unused_crate_dependencies = "deny"
+rust.unused_must_use = "deny"
+rustdoc.all = "warn"
 
 [dependencies]
 risc0-groth16 = "1.2.3"
 risc0-binfmt = "1.2.3"
 risc0-circuit-recursion = "1.2.3"
 risc0-zkp = "1.2.3"
-risc0-zkvm = "1.2.3"
 zkaleido.workspace = true
 hex.workspace = true
-
-sha2.workspace = true
 
 [dev-dependencies]
 hex.workspace = true
